roomSchedulingService/requirements.txt
blinker==1.9.0
certifi==2025.10.5
charset-normalizer
click
colorama==0.4.6
dotenv==0.9.9
Flask==3.1.2
Flask-MySQLdb==2.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python
mysqlclient==2.2.7
python-dotenv==1.2.1
requests==2.32.5
urllib3==2.5.0
Werkzeug==3.1.3
